Clean up examples.
