--- examples/coap/README.md
@@ -1,1 +1,0 @@
-Coap support is a work in progress.
