--- examples/coap/observe_client.js
@@ -1,10 +1,0 @@
-const coap  = require('coap')
-    , req   = coap.request({
-                observe: true
-              })
-
-req.on('response', function(res) {
-  res.pipe(process.stdout)
-})
-
-req.end()
